# services/authentication-service/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RFx Sourcing — Authentication Service
  description: |
    Issues HS256 JWTs consumed by the RFx Sourcing API Gateway and downstream services.
    All error responses follow RFC 7807 Problem Details (`application/problem+json`).
  version: "0.1.0"
  contact:
    name: Ramkumar (Solution Architect)
servers:
  - url: http://localhost:5003/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          example: eyJhbGciOiJIUzI1NiIs...

    # ── Responses ─────────────────────────────
...
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        tokenType:
          type: string
          enum: [Bearer]
        accessTokenExpiresInSeconds:
          type: integer
          example: 900
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details object (`application/problem+json`).
      properties:
        type:
          type: string
          format: uri
          example: https://rfx.gep.local/probs/invalid-credentials
        title:
          type: string
          example: Invalid credentials
        status:
          type: integer
          example: 401
        detail:
          type: string
          example: Username or password is incorrect.
        instance:
          type: string
          example: /api/v1/auth/login
        correlationId:
          type: string
          format: uuid
          example: abc-123-def-456

# ──────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange refresh token for a new access token
      description: |
        Supply a valid refresh token to obtain a new access token.
        The refresh token itself is **not** rotated (same token reused until expiry).
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Validation error — missing or empty refreshToken field
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Refresh token invalid, malformed, or expired
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://rfx.gep.local/probs/invalid-refresh-token
                title: Invalid or expired refresh token
                status: 401
                detail: The supplied refresh token is invalid, malformed, or has expired.
                instance: /api/v1/auth/refresh
                correlationId: abc-125
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...